phase3_chatbot/backend/requirements.txt
fastapi
uvicorn
cohere
sqlalchemy
python-dotenv
psycopg2-binary
requests